requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
dj-database-url==0.5.0
Django==4.0.6
django-environ==0.9.0
django-heroku==0.3.1
gunicorn==20.1.0
Pillow==9.2.0
psycopg2==2.9.3
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0
